x/todo/keeper/msg_server_paper_test.go
package keeper_test

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"vbi-cosmos-basic/x/todo/types"
)

func TestPaperMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreatePaper(ctx, &types.MsgCreatePaper{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestPaperMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdatePaper
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdatePaper{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdatePaper{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdatePaper{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreatePaper(ctx, &types.MsgCreatePaper{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdatePaper(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestPaperMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeletePaper
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeletePaper{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeletePaper{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeletePaper{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreatePaper(ctx, &types.MsgCreatePaper{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeletePaper(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
